phan: Improve phpunit stub munging (#37164)
In Psalm, `@psalm-assert Foo $arg` asserts both that if the function
doesn't throw then the argument is an instance of Foo _and_ that if the
function does throw then the argument is not an instance of Foo. If you
want only the first without the second, you do `@psalm-assert =Foo $arg`
instead.

Phan, on the other hand, never makes the inverse assumption. So
effectively `@phan-assert Foo $arg` is equivalent to
`@psalm-assert =Foo $arg`, and trying to `@phan-assert =Foo` is a syntax
error.

Thus, when we're munging the phpunit annotations, the `=` needs to be
stripped when converting `@psalm-assert` to `@phan-assert`. This will
allow Phan to infer correct types for `assertInstanceOf()` and
`assertSame()`.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/8914715128

Upstream-Ref: Automattic/jetpack@8b9fb8454352f62f17672789f4f49cbd7a459f95